// src/api/rathena/DatabaseDriver.types.ts
export type CartInventoryEntity = zod.infer<typeof CartInventoryEntityType>;
export const CartInventoryEntityType = zod.object({
  amount: zod.number(),
  attribute: zod.number(),
  bound: zod.number(),
  card0: zod.number(),
  card1: zod.number(),
  card2: zod.number(),
  card3: zod.number(),
  char_id: zod.number(),
  enchantgrade: zod.number(),
  equip: zod.number(),
  expire_time: zod.number(),
  id: zod.number(),
  identify: zod.number(),
  nameid: zod.number(),
  option_id0: zod.number(),
  option_id1: zod.number(),
  option_id2: zod.number(),
  option_id3: zod.number(),
  option_id4: zod.number(),
  option_parm0: zod.number(),
  option_parm1: zod.number(),
  option_parm2: zod.number(),
  option_parm3: zod.number(),
  option_parm4: zod.number(),
  option_val0: zod.number(),
  option_val1: zod.number(),
  option_val2: zod.number(),
  option_val3: zod.number(),
  option_val4: zod.number(),
  refine: zod.number(),
  unique_id: zod.string(),
});
// ...
export type InventoryEntity = zod.infer<typeof InventoryEntityType>;
export const InventoryEntityType = zod.object({
  amount: zod.number(),
  attribute: zod.number(),
  bound: zod.number(),
  card0: zod.number(),
  card1: zod.number(),
  card2: zod.number(),
  card3: zod.number(),
  char_id: zod.number(),
  enchantgrade: zod.number(),
  equip: zod.number(),
  equip_switch: zod.number(),
  expire_time: zod.number(),
  favorite: zod.number(),
  id: zod.number(),
  identify: zod.number(),
  nameid: zod.number(),
  option_id0: zod.number(),
  option_id1: zod.number(),
  option_id2: zod.number(),
  option_id3: zod.number(),
  option_id4: zod.number(),
  option_parm0: zod.number(),
  option_parm1: zod.number(),
  option_parm2: zod.number(),
  option_parm3: zod.number(),
  option_parm4: zod.number(),
  option_val0: zod.number(),
  option_val1: zod.number(),
  option_val2: zod.number(),
  option_val3: zod.number(),
  option_val4: zod.number(),
  refine: zod.number(),
  unique_id: zod.string(),
});